requirements.txt
# Core Framework Dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database Connections
neo4j>=5.15.0
qdrant-client>=1.7.0

# LangChain and LLM Integration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-groq>=1.0.0
groq>=0.4.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-multipart>=0.0.6
markdown>=3.5.0

# Local Embedding Models
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.36.0

# Data Processing and Utilities
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Visualization
plotly>=5.17.0

# HTTP and API
httpx>=0.25.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
pylint>=3.0.0

# Jupyter for Exploration
jupyter>=1.0.0
ipykernel>=6.26.0